build: presets, warnings-as-signal, embeddability fixes
- CMakePresets.json: dev (RelWithDebInfo -O2 -g, asserts on), release
  (installs to ~/projects/dist/strikeengine), asan (ASan+UBSan), each with
  configure/build/test presets.
- -Wall -Wextra on the engine; the 3 warnings it found are fixed (unused
  wingSweepDeg, sign-compare on lockedTargetId, and three unused test
  variables). The library is now warning-clean.
- STRIKEENGINE_BUILD_TESTS option so embedders can build the engine without
  its test suite (StrikeSim's dev preset uses this).
- CONFIGURE_DEPENDS on the source glob: new .cpp files are picked up without
  a manual reconfigure.
- CMAKE_SOURCE_DIR -> CMAKE_CURRENT_SOURCE_DIR for include dirs, shader
  paths and test source dirs, so add_subdirectory() consumers get the right
  paths.
- Vulkan shader compiles with the Vulkan::glslc imported target (matching
  the sim) instead of a bare glslc from PATH.
- Explicit Debug default for bare configures (presets carry the optimized
  configurations), and a 900 s CTest timeout.

Dev suite: 45/45 in 0.6 s (was ~8 s unoptimized).

+++ src/strikeengine/kernel/SimulationKernel.cpp
@@ -1680,8 +1680,7 @@ namespace StrikeEngine::Kernel {
             bool locked = parent < seekerBlock.isLocked.size() && seekerBlock.isLocked[parent];
             if (locked && spec.targetIndex >= 0) {
                 locked = (parent < seekerBlock.lockedTargetId.size() &&
-                          seekerBlock.lockedTargetId[parent] ==
-                              static_cast<std::int64_t>(tgt));
+                          seekerBlock.lockedTargetId[parent] == tgt);
             }
             if (!locked) {
                 pl.lockSince = -1.0;
